pyproject.toml
[project]
name = "docx-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
